refactor: 単一利用モジュールを lib から各 bin 配下へ移動 (#388) (#404)
## 概要

[PR #403](https://github.com/toshiki670/dotfiles/pull/403)（Epic →
main）の**レビュー指摘対応**。マージ先は Epic 統合ブランチ。

> - `src/lint.rs`: dotfiles-lint でしか使われていないので Lib で共通化は不要 →
`src/bin/dotfiles-lint` へ
> - `src/gcm.rs`: gcm でしか使われていないので Lib で共通化は不要 → `src/bin/gcm` へ

1 つの bin でしか使わないロジックは lib に置かず、その bin 配下へ移す方針に統一する。

## 変更内容

- `src/lint.rs` + `src/lint/{classify,collect}.rs` →
**`src/bin/dotfiles-lint/lint{,/...}`**（dotfiles-lint 専用）。
- `src/gcm.rs` → **`src/bin/gcm/proposals.rs`**（gcm 専用）。
- 同じ原則を適用し、**`remote_exists` も lib → `src/bin/git-upstream`**
へ移動（git-upstream 専用）。指摘は 2 件でしたが、同一の単一利用ケースのため一貫性のため併せて移しました（不要なら戻せます）。
- **lib には複数 bin（`v-sync` / `daily-check-worker` / `gh-clone`）で共有する
`command_exists` のみ**残す。
- 各 `main.rs` の `use dotfiles::...` を bin 内 `mod`
参照へ変更。テストも各モジュールと一緒に移動。

## 検証（ローカル）

- `cargo fmt --check` / `clippy -D warnings`（**dead_code 警告なし** = 移動した
`pub` 項目はすべて各 bin 内で使用）/ `cargo test`（**19**、各 bin のテストバイナリに分散）。
- 各 bin スモーク（git-upstream / gcm / dotfiles-lint）、`dotfiles-lint check` →
EXIT=0。

Refs #388, #403
