+++ src/bin/git-upstream/main.rs
@@ -60,9 +60,14 @@ fn git_available() -> bool {
 /// upstream リモートが登録済みか判定する。
 fn has_upstream() -> bool {
     match Command::new("git").arg("remote").output() {
-        Ok(out) => dotfiles::remote_exists(&String::from_utf8_lossy(&out.stdout), REPOSITORY),
+        Ok(out) => remote_exists(&String::from_utf8_lossy(&out.stdout), REPOSITORY),
         Err(_) => false,
     }
+}
+
+/// `git remote` の出力（1 行 1 リモート名）に指定したリモートが含まれるか判定する。
+fn remote_exists(remotes_output: &str, name: &str) -> bool {
+    remotes_output.lines().any(|line| line.trim() == name)
 }
 
 /// `git <args>` を stdio 継承で実行し、終了コードを返す。
@@ -103,4 +108,31 @@ fn initialize(remote_url: &str) -> i32 {
     }
 
     run(&["fetch", REPOSITORY])
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn remote_exists_finds_named_remote() {
+        let output = "origin\nupstream\n";
+        assert!(remote_exists(output, "upstream"));
+        assert!(remote_exists(output, "origin"));
+    }
+
+    #[test]
+    fn remote_exists_returns_false_when_absent() {
+        assert!(!remote_exists("origin\n", "upstream"));
+    }
+
+    #[test]
+    fn remote_exists_ignores_surrounding_whitespace() {
+        assert!(remote_exists("  upstream  \n", "upstream"));
+    }
+
+    #[test]
+    fn remote_exists_is_exact_match() {
+        assert!(!remote_exists("upstream-mirror\n", "upstream"));
+    }
 }
